[svn] VERY IMPORTANT: - si.c was using "random()" instead of new tsi_random() (for layers setup). - math_random.c was missing a cast to double on "tsi_random_real()".
others:
- tsi_io.c replaced strncmp with strncasecmp
- fixed get COLLECTIVE_COMPARE config in tsi.c
- fixed timeval -> my_time (timer type abstraction)



THERE IS STILL ONE BUG IN THE CORRELATION.

+++ src/math_random.c
@@ -206,7 +206,7 @@ unsigned long tsi_random(void)
 	//TODO: try int28()
 	return genrand_int31();
 #else
-    return random();
+    return (unsigned long)random();
 #endif
 }
 
@@ -215,6 +215,6 @@ double tsi_random_real(void)
 #ifdef NEW_RAND
 	return genrand_real2();
 #else
-    return random() / RAND_MAX;
+    return ((double)random()) / RAND_MAX;
 #endif
 }
